requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
psycopg2-binary==2.9.9
Werkzeug==3.0.3
Flask
psycopg2-binary
python-dotenv
flask_cors==3.0.10
gunicorn==20.1.0
flask_cors